refactor: rename NnsNeuronMetaInfoCard
