+++ frontend/src/lib/components/neuron-detail/NnsNeuronMetaInfoCard.svelte
